Moved pages/views to corresponding folders
