+++ src/components/views/LoginVIew/LoginPage.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useContext, FC, useEffect } from "react";
 import { Box, Typography, TextField, Button, Paper, Grid } from "@mui/material";
-import { TheContext } from "../TheContext";
-import CreateAccount from "../components/CreateAccount";
-import BACKEND_URL from "../backendUrl";
+import { TheContext } from "../../../TheContext";
+import CreateAccount from "../CreateAccountVIew/CreateAccount";
+import BACKEND_URL from "../../../backendUrl";
 import {
   loginButton,
   loginBox,
@@ -10,7 +10,7 @@ import {
   loginHeaderTitleText,
   loginHeaderContentText,
   loginPaper
-} from "../sxStyles";
+} from "../../../sxStyles";
 
 interface IProps {
   logged: boolean;
